requirements.txt
torch>=1.9.0
torchvision>=0.10.0
visdom>=0.2.0
numpy>=1.19.5
Pillow>=8.3.1
tqdm>=4.62.0
tensorboard>=2.4.0
matplotlib>=3.4.0